Move Discord.Monad to Discord.Internal.Monad to reduce overuse risk
